fix(math): Added missing functions and fixed some compile and docs issues
Core changes:
- Added templated Abs, Max, Min to utils.hpp
- Added missing operators to vector3d.hpp
- Added missing operators to vector2d.hpp
- Added ToDouble() for fxp
- Changed ToFloat() to give float instead of double
- Fixed ToFloat(), ToDouble(), ToInt() not being const
- Fixed documentation errors
- Removed stdexcept include from matrix matrix_stack.hpp

+++ impl/plane.hpp
@@ -24,7 +24,7 @@ namespace SaturnMath
         Fxp d;          /**< Signed distance from origin to plane */
 
         /** @brief Default constructor. Creates XY plane at origin. */
-        constexpr Plane() : normal(Vector3D::UnitZ()), d(0) {}
+        constexpr Plane() : normal(Vector3D::UnitZ()), d((int16_t)0) {}
 
         /**
          * @brief Creates plane from normal and distance.
